Final secure version.
